Keep the MJCF kinematics out of the placo module
`kinematics.py` imports `placo` at module scope, and only the `hardware` extra
carries it. Putting the MJCF solver there gave the `yam` extra an undeclared
dependency on another one: `uv run --locked --exact --extra yam` could no
longer import the YAM driver at all. It has its own module now, and the two
files say why they are two.

The site is checked: `mj_name2id` answers -1 for a name the model does not
carry, and indexing with it reads the model's last site, so the arm would solve
for another frame.

`reach_postures` takes the whole target rather than its x and y. Where a seed
comes from is the arm's business, and the shared class has no reason to say it
is planar. The tolerances are the class's own constants.
